+++ src/components/dataTable.js
@@ -11,7 +11,6 @@
       GetMe,
       useText,
       ModelProvider,
-      useEndpoint,
       useAllQuery,
       useFilter,
     } = B;
@@ -305,36 +304,13 @@
       setSearch(event.target.value);
     };
 
-    const isFlatValue = value =>
-      typeof value === 'string' ||
-      typeof value === 'number' ||
-      typeof value === 'boolean';
-
     const history = isDev ? {} : useHistory();
 
-    const handleRowClick = rowValue => {
+    const handleRowClick = (endpoint, context) => {
       if (isDev) return;
-      B.triggerEvent('OnRowClick', rowValue);
+      B.triggerEvent('OnRowClick', endpoint, context);
       if (hasLink) {
-        const { id, params } = linkTo;
-        const newParams = Object.entries(params).reduce((acc, cv) => {
-          const key = cv[0];
-          const value = cv[1];
-          if (isFlatValue(value[0])) {
-            acc[key] = value;
-          } else {
-            const propId = value[0].id;
-            const property = getProperty(propId).name;
-            acc[key] = [rowValue[property].toString()];
-          }
-          return acc;
-        }, {});
-
-        const endpointParams = {
-          id,
-          params: newParams,
-        };
-        history.push(useEndpoint(endpointParams));
+        history.push(endpoint);
       }
     };
 
@@ -370,14 +346,23 @@
 
       const rows = results.map(value => (
         <ModelProvider value={value} id={model}>
-          <TableRow
-            key={value[0]}
-            classes={{ root: classes.bodyRow }}
-            onClick={() => handleRowClick(value)}
-            data-id={value.id}
-          >
-            <B.InteractionScope>{children}</B.InteractionScope>
-          </TableRow>
+          <B.InteractionScope model={model}>
+            {context => (
+              <TableRow
+                key={value[0]}
+                classes={{ root: classes.bodyRow }}
+                data-id={value.id}
+              >
+                <Children
+                  linkTo={linkTo}
+                  handleRowClick={handleRowClick}
+                  context={context}
+                >
+                  {children}
+                </Children>
+              </TableRow>
+            )}
+          </B.InteractionScope>
         </ModelProvider>
       ));
 
@@ -389,19 +374,21 @@
     };
 
     const renderTableContent = () => {
-      let tableContent = Array.from(Array(amountOfRows).keys()).map(idx => (
-        <TableRow key={idx} classes={{ root: classes.bodyRow }}>
-          <B.InteractionScope>{children}</B.InteractionScope>
-        </TableRow>
-      ));
       if (isDev) {
-        tableContent = (
+        return (
           <TableRow classes={{ root: classes.bodyRow }}>{children}</TableRow>
         );
-      } else if (model) {
-        tableContent = tableContentModel();
       }
-      return tableContent;
+
+      if (model) {
+        return tableContentModel();
+      }
+
+      return Array.from(Array(amountOfRows).keys()).map(idx => (
+        <TableRow key={idx} classes={{ root: classes.bodyRow }}>
+          {children}
+        </TableRow>
+      ));
     };
 
     useEffect(() => {